+++ database/migrations/2020_08_03_080014_create_users_permissions_table.php
@@ -15,8 +15,8 @@ class CreateUsersPermissionsTable extends Migration
     {
         Schema::create('users_permissions', function (Blueprint $table) {
             $table->id();
-            $table->unsignedInteger('userID');
-            $table->unsignedInteger('permissonID');
+            $table->unsignedBigInteger('userID');
+            $table->unsignedBigInteger('permissonID');
             $table->timestamps();
 
             $table->foreign('userID')
